+++ R/pq_freq.R
@@ -74,7 +74,7 @@ pq1_to_freq = function(dat, freq) {
 
 #' converting frequency of daily data
 #' 
-#' \code{pq_to_freq} convert a daily OHLC dataframe into a specified frequency.
+#' \code{pq_freq} convert a daily OHLC dataframe into a specified frequency.
 #' 
 #' @param dt a list/dataframe of time series dataset.
 #' @param freq the frequency that the input daily data will converted to. It supports weekly, monthly, quarterly and yearly.
@@ -84,12 +84,12 @@ pq1_to_freq = function(dat, freq) {
 #' \donttest{
 #' dts = md_stock(c("^000001", "000001"), date_range = 'max', source = '163')
 #' 
-#' dts_weekly = pq_to_freq(dts, "weekly")
+#' dts_weekly = pq_freq(dts, "weekly")
 #' }
 #' 
 #' @export
 #' 
-pq_to_freq = function(dt, freq, print_step=1L) {
+pq_freq = function(dt, freq, print_step=1L) {
     symbol = NULL
     # check freq argument
     freq = check_arg(freq, c("weekly","monthly","quarterly","yearly"))